+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/CustomerSheet/CustomerSheet.swift
@@ -164,10 +164,17 @@ public class CustomerSheet {
 extension CustomerSheet {
     func loadPaymentMethodInfo(completion: @escaping (Result<([STPPaymentMethod], CustomerPaymentOption?, STPElementsSession), Error>) -> Void) {
         Task {
+            if let paymentMethodTypes = self.customerAdapter.paymentMethodTypes {
+                if case .failure(let error) = paymentMethodTypes.customerSheetSupportedPaymentMethodTypes(CustomerSheet.supportedPaymentMethods) {
+                    completion(.failure(error))
+                    return
+                }
+            }
+
             do {
                 async let paymentMethodsResult = try customerAdapter.fetchPaymentMethods()
                 async let selectedPaymentMethodResult = try self.customerAdapter.fetchSelectedPaymentOption()
-                async let elementsSessionResult = try self.configuration.apiClient.retrieveElementsSessionForCustomerSheet()
+                async let elementsSessionResult = try self.configuration.apiClient.retrieveElementsSessionForCustomerSheet(paymentMethodTypes: self.customerAdapter.paymentMethodTypes)
 
                 // Ensure local specs are loaded prior to the ones from elementSession
                 await loadFormSpecs()